requirements_full.txt
# Core Dependencies
tensorflow>=2.10.0,<2.16.0
numpy>=1.23.0,<2.0.0
pandas>=1.5.0,<3.0.0
scikit-learn>=1.2.0,<2.0.0

# Web Application
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
scipy>=1.10.0

# Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Database
# SQLite is built-in with Python
# Additional dependencies for authentication
bcrypt>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
